Peer: Removed extended warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@23990 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 4ec49cd3a650525bd6d495b2c6ec5a8fdf39edfb

+++ lsop/de.uniol.inf.is.odysseus.sports.sportsql/src/de/uniol/inf/is/odysseus/sports/sportsql/physicaloperator/SportsHeatMapPO.java
@@ -98,8 +98,8 @@ public class SportsHeatMapPO<T extends Tuple<?>> extends AbstractPipe<T, Tuple> 
 		int newX = (int) (xValue - minX);
 		int newY = (int) (yValue - minY);
 
-		int xArray = (int) (newX / pixelsPerSquareHorizontal);
-		int yArray = (int) (newY / pixelsPerSquareVertical);
+		int xArray = newX / pixelsPerSquareHorizontal;
+		int yArray = newY / pixelsPerSquareVertical;
 
 		// If the players leave the field (I guess)
 		if (xArray >= numTilesX)
